requirements.txt
asyncio-nats-client==0.11.5
asyncio-nats-streaming==0.4.0
protobuf==4.21.6
